Basedpyright integration, too much to list....
[Addition] Basedpyright is now being used to make sure the project is the most type safe it can be
[Addition] Added unit tests for worker
[Addition] Added OptionsManager.getLaunchParametersList()
[Addition] Added proton OptionKeys for later
[Addition] Finished launch parameter code
[Refactor] Many functions have been stripped of returning multiple types
[Refactor] Constant IS_SCRIPT is now IS_DEBUG, it has been turned into a literal boolean
[Refactor] imports have been adjusted throughout the program to reduce file size and replace deprecated type hints
[Refactor] @staticmethod has been given to ModTypes.all_types()
[Refactor] Fixed some cyclical importing
[Bugfix] Added slot decorator to several functions
[Refactor] Improved some unit tests
[Refactor] Updated patool to 4.0.5, pyinstaller to 6.22.2, pytest to 9.1.1 and PySide6 to 6.10.3
[Refactor] Changed PySide6 package to use the essentials wheel instead of the default to save storage
[Refactor] venv directory name has been changed to ".venv"
[Refactor] Added clarification about the project and anti-virus false positives in README
[Refactor] venv.bat and venv.sh have been simplified
[Refactor] Bumped to python 3.12
[Removed] semanic_versioning and send2trash, Qt has features to replace these
[Bugfix] Deleting a disabled mod no longer deletes the disabled mods folder. I tried to see how long this bug has been in the program and it appears to be at least since 1.2 wow, I added unit tests to prevent that bug specifically from happening again
[Removed] Removed permissionCheck()
[Removed] isPrerelase() and taking prerelease versions into account
[Removed] Some type hints have been removed to save line space on the screen and with Basedpyright the type hints should inline

+++ src/objects/worker.py
@@ -1,13 +1,7 @@
 import logging
-import os
 import shutil
-from collections.abc import Callable
-from typing import Any
 
-from PySide6.QtCore import QCoreApplication as qapp
-from PySide6.QtCore import QObject, Signal, QMutex, QMutexLocker
-
-import src.helpers.helper as helper
+from PySide6.QtCore import QObject, Signal, QMutex, QMutexLocker, QFileInfo, QCoreApplication as qapp
 
 class Worker(QObject):
     setTotalProgress = Signal(int)
@@ -24,13 +18,13 @@ class Worker(QObject):
 
     cancel = False
 
-    mutex: QMutex = None # Should be set externally by the ProgressWidget class
+    mutex: QMutex = QMutex() # Should be set externally by the ProgressWidget class
 
     def __init__(self) -> None:
         super().__init__()
         logging.getLogger(__name__)
 
-    def start() -> None:
+    def start(self) -> None:
         ...
 
     def onCancel(self) -> None:
@@ -54,48 +48,31 @@ class Worker(QObject):
     def move(self, src: str, dest: str) -> None:
         '''`shutil.move()` with some extra exception handling'''
 
-        # Overwrite mod
-        if os.path.exists(dest):
-            shutil.rmtree(dest, onerror=self.onError)
+        dest_file_info = QFileInfo(dest)
+        src_file_info = QFileInfo(src)
 
         # Will try to move the file, if there is an exception, fix the issue and try again
         try:
+
+            if not QFileInfo(dest_file_info.absolutePath()).exists():
+                raise FileNotFoundError(f"File's destination does not exist\n{dest_file_info.absolutePath()}")
+
+            # Overwrite mod
+            if dest_file_info.exists():
+                shutil.rmtree(dest)
+            
+            if not src_file_info.exists():
+                raise FileNotFoundError(f"File to be moved does not exist\n{src_file_info.absoluteFilePath()}")
+
             shutil.move(src, dest)
             logging.info('Moved file %s to destination %s', src, dest)
 
-        except PermissionError:
-            
-            # Grab all files in mod
-            for root, dirs, files in os.walk(src):
+        except Exception as e:
 
-                self.addTotalProgress.emit(2 + len(dirs) + len(files))
-                
-                # Checking files for perm errors
-                for file in files:
-                    self.setCurrentProgress.emit(1, qapp.translate('Worker', 'Checking file permissions of') + f' {file}')
-                    file_path = os.path.join(root, file)
-                    helper.permissionCheck(file_path)
-                
-                # Checking folders for perm errors
-                for dir in dirs:
-                    self.setCurrentProgress.emit(1, qapp.translate('Worker', 'Checking folder permissions of') + f' {dir}')
-                    dir_path = os.path.join(root, dir)
-                    helper.permissionCheck(dir_path)
-                
-                # Checking mod directory for perm errors
-                self.setCurrentProgress.emit(1, qapp.translate('Worker', 'Checking folder permissions of') + f' {root}')
-                helper.permissionCheck(root)
-
-            self.setCurrentProgress.emit(1, qapp.translate('Worker', 'Fixing install for') + f' {os.path.basename(src)}')
             # If shutil.move made a partial dir of the mod delete it
-            if os.path.exists(dest):
-                shutil.rmtree(dest, onerror=self.onError)
-
-    def onError(self, func: Callable[[Any], Any], path: str, exc_info: int) -> None:
-        """Used for `shutil.rmtree()`s `onerror` kwarg"""
-
-        logging.warning('An error was raised in shutil:\n%s', exc_info)
-
-        if not helper.permissionCheck(path):
-
-            func(path)
+            if dest_file_info.exists():
+                shutil.rmtree(dest)
+            
+            logging.warning('An error was raised in shutil:\n%s', e)
+            self.error.emit(f"{qapp.translate("Worker", "Error")}: {e}")
+
